CF-131 smooth scroll to tab when out of view
