+++ frontend/src/pages/CourseSelector/courseTabs/CourseTab.jsx
@@ -1,0 +1,74 @@
+import React, { useEffect, useRef, useState } from "react";
+import { Button } from "antd";
+import { useSelector, useDispatch } from "react-redux";
+import { courseTabActions } from "../../../actions/courseTabActions";
+import "./courseTabs.less";
+import { Draggable } from "react-beautiful-dnd";
+import { CloseOutlined } from "@ant-design/icons";
+import { useIntersectionObserver } from "../../../hooks/useIntersectionObserver";
+
+const CourseTab = ({ tab, index }) => {
+  const [scrolledTo, setScrolledTo] = useState(false);
+  const dispatch = useDispatch();
+  const ref = useRef(null);
+  const tabInView = useIntersectionObserver(ref);
+  const { active } = useSelector((state) => state.tabs);
+
+  const getDraggableStyle = (style, snapshot) => {
+    // lock x axis when dragging
+    if (style.transform) {
+      const axisLockX = `${style.transform.split(",").shift()}, 0px)`;
+      return {
+        ...style,
+        transform: axisLockX,
+      };
+    }
+    return style;
+  };
+
+  useEffect(() => {
+    if (active === index && !scrolledTo) {
+      ref.current.scrollIntoView({
+        behavior: "smooth",
+        block: "start",
+        inline: "nearest",
+      });
+      setScrolledTo(true);
+    }
+    setScrolledTo(false);
+  }, [active, tabInView, index, scrolledTo]);
+
+  return (
+    <Draggable key={tab} draggableId={tab} index={index}>
+      {(draggableProvided, draggableSnapshot) => (
+        <div
+          className={index === active ? "cs-tab active" : "cs-tab"}
+          onClick={() => dispatch(courseTabActions("SET_ACTIVE_TAB", index))}
+          ref={(r) => {
+            ref.current = r;
+            draggableProvided.innerRef(r);
+          }}
+          {...draggableProvided.draggableProps}
+          {...draggableProvided.dragHandleProps}
+          style={getDraggableStyle(
+            draggableProvided.draggableProps.style,
+            draggableSnapshot
+          )}
+        >
+          <span className="cs-tab-name">{tab}</span>
+          <Button
+            type="text"
+            size="small"
+            icon={<CloseOutlined style={{ fontSize: "12px" }} />}
+            onClick={(e) => {
+              e.stopPropagation(); // stop propagation for above tab onclick event
+              dispatch(courseTabActions("REMOVE_TAB", index));
+            }}
+          />
+        </div>
+      )}
+    </Draggable>
+  );
+};
+
+export default CourseTab;
