+++ frontend/src/pages/CourseSelector/courseTabs/courseTabs.less
@@ -17,6 +17,9 @@
   min-height: 55px;
   height: fit-content;
   padding-top: 15px;
+  &::-webkit-scrollbar {
+    display: none;
+  }
 
   .cs-tab {
     border-radius: 10px 0 0 0 !important;
@@ -25,7 +28,7 @@
     justify-content: center;
     align-items: center;
     gap: 5px;
-    padding: 8px 6px 8px 16px;
+    padding: 6px 6px 6px 16px;
     margin-left: 2px;
     border: 1px solid #f0f0f0;
     transition: color 0.3s cubic-bezier(0.645, 0.045, 0.355, 1);
