+++ frontend/src/pages/CourseSelector/main.jsx
@@ -4,7 +4,7 @@ import { Switch } from "antd";
 import { useSelector } from "react-redux";
 import CourseMenu from "./courseMenu/CourseMenu";
 import CourseDescription from "./courseDescription/CourseDescription";
-import { CourseTabs } from "./CourseTabs";
+import { CourseTabs } from "./courseTabs/CourseTabs";
 import "./main.less";
 import SearchCourse from "./SearchCourse";
 
@@ -43,18 +43,19 @@ export default function CourseSelector() {
           )}
         </div>
         <SearchCourse />
-
-        <Switch 
+        <Switch
           className="cs-toggle-locked"
-          onChange={() => setShowLockedCourses(prev => !prev)}
+          onChange={() => setShowLockedCourses((prev) => !prev)}
           checkedChildren="locked courses shown"
           unCheckedChildren="locked courses hidden"
         />
-
       </div>
       <CourseTabs />
       <div className="cs-bottom-cont">
-        <CourseMenu structure={structure} showLockedCourses={showLockedCourses} />
+        <CourseMenu
+          structure={structure}
+          showLockedCourses={showLockedCourses}
+        />
         <CourseDescription structure={structure} />
       </div>
     </div>
